Remake graph for de_expulsion_final (#25)
Fix goal node for de_dust2_middle
Also the map was completely covered with waypoints.
Upload fixed graphs
Remake graph for de_expulsion_final
The previous graph had the goal node in the wrong place. Because of this, the bots could not plant the bomb. I redid the graph, put the goal nodes in the right places, completely covered the map with nodes, including the roofs where the bomb site is also located.
Fix typo in readme